pyproject.toml
[project]
name = "scene-builder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.120.1",
    "google-generativeai>=0.8.5",
    "httpx>=0.28.1",
    "pydantic>=2.12.3",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.38.0",
]
